working to refactor main data import function
