+++ client/api.js
@@ -1,7 +1,7 @@
 import request from 'superagent'
 
-export function retrieveRandomTitle () {
-  return request.get('/v1/books/review')
+export function retrieveAllTitles () {
+  return request.get('/v1/books/alllist')
     .then(resp => {
       console.log('api.js: ' + resp.body)
       const bookData = resp.body
@@ -11,15 +11,3 @@ export function retrieveRandomTitle () {
       throw Error('Requires /v1/books/review api route')
     })
 }
-
-export function retrieveAllTitles () {
-  return request.get('/v1/books/alllist')
-    .then(resp => {
-      const titles = resp.body
-      return titles
-    })
-    .catch(() => {
-      throw Error('Requires /v1/alllist api route')
-    })
-}
-
