+++ routes/books.js
@@ -2,12 +2,7 @@ const express = require('express')
 const db = require('../db')
 const router = express.Router()
 
-router.get('/review', (req, res) => {
-  db.retrieveRandomTitle()
-    .then(bookData => res.send(bookData))
-})
-
-router.get('/alllist', (req, res) => {
+router.get('/', (req, res) => {
   db.retrieveAllTitles()
     .then(bookData => res.send(bookData))
 })
